web/requirements.txt
fastapi
uvicorn[standard]
jinja2
httpx
python-dotenv
itsdangerous
python-multipart
sqlalchemy[asyncio]
alembic
aiosqlite
